+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.43.0](https://github.com/serlo/api.serlo.org/compare/v0.42.1..v0.43.0) - July 12, 2022
+
+### Added
+
+- Add `user.addRole` mutation #663
+
+- Add `user.removeRole` mutation #663
+
 ## [v0.42.1](https://github.com/serlo/api.serlo.org/compare/v0.42.0..v0.42.1) - July 6, 2022
 
 ### Added